+++ android/library/src/main/res/values/attrs.xml
@@ -282,7 +282,4 @@
 		<attr name="classPK"/>
 	</declare-styleable>
 
-	<declare-styleable name="GetUserScreenlet">
-		<attr format="string" name="getUserBy"/>
-	</declare-styleable>
 </resources>
