+++ android/samples/getuserscreenlet/src/main/java/com/liferay/mobile/screens/user/GetUserListener.java
