+++ android/samples/getuserscreenlet/src/main/java/com/liferay/mobile/screens/user/interactor/GetUserInteractor.java
