+++ android/samples/getuserscreenlet/src/main/res/layout/custom_get_user_row.xml
